Endless Terrain
Planes can now be generated around a given transform. Each chunk has a flag to say whether it is visible or not, depending on its distance to the player object (which is just a cube for now) . There is a dictionary to keep track of all of the chunks visited in the game, and a list to keep tarck of which chunks were visited since the last update loop.