Collision Refactoring
The mesh collider is now of a chunk with a lower LOD than the current chunk. On top of this, only the mesh collider of the current chunk is generated. This speeds up the game as only one collider mesh has to be generated and it is of a lower resolution than the mesh that the player is currently on.