Water Shader and Procedural Generation
Finally the two worlds meet (although I am cheating a bit) After many hours of scratching my head, I was able to add the shader on a plane as a prefab to each chunk. This process was slow and EXTREMELY expensive on the GPU (unity crashed for me, and I am running a GTX 1070) So I spent another hour thinking of how to optimise. I decided on having just one plane (albeit a big one, measuring up at 480x480 units) and have it follow the player. This turned out to be a much more efficient way, whilst still acheiving the same effect.

I am super happy with the results of this!